Move core views, urls, tests into home app
