+++ src/rf/rf/urls.py
@@ -8,7 +8,6 @@ from django.contrib import admin
 from rest_framework import routers
 
 import apps.home.urls
-import apps.core.urls
 import apps.user.urls
 
 admin.autodiscover()
@@ -16,7 +15,6 @@ admin.autodiscover()
 urlpatterns = patterns(
     '',
     url(r'^', include(apps.home.urls)),
-    url(r'^', include(apps.core.urls)),
     url(r'^user/', include(apps.user.urls)),
     url(r'^admin/', include(admin.site.urls)),
 )
